+++ tests/NClient/NClient.Tests/ClientTests/ResponseClientTest.cs
@@ -92,7 +92,7 @@ namespace NClient.Tests.ClientTests
             result.Should().NotBeNull();
             using var assertionScope = new AssertionScope();
             result.Data.Should().Be(id);
-            result.StatusCode.Should().Be(HttpStatusCode.OK);
+            result.StatusCode.Should().Be((int)HttpStatusCode.OK);
         }
 
         [Test]
@@ -105,7 +105,7 @@ namespace NClient.Tests.ClientTests
 
             result.Should().NotBeNull();
             using var assertionScope = new AssertionScope();
-            result.StatusCode.Should().Be(HttpStatusCode.BadRequest);
+            result.StatusCode.Should().Be((int)HttpStatusCode.BadRequest);
         }
 
         [Test]
@@ -118,7 +118,7 @@ namespace NClient.Tests.ClientTests
 
             result.Should().NotBeNull();
             using var assertionScope = new AssertionScope();
-            result.StatusCode.Should().Be(HttpStatusCode.BadRequest);
+            result.StatusCode.Should().Be((int)HttpStatusCode.BadRequest);
             result.Content.ToString().Should().Be(JsonSerializer.Serialize(BadRequestError));
         }
 
@@ -132,7 +132,7 @@ namespace NClient.Tests.ClientTests
 
             result.Should().NotBeNull();
             using var assertionScope = new AssertionScope();
-            result.StatusCode.Should().Be(HttpStatusCode.InternalServerError);
+            result.StatusCode.Should().Be((int)HttpStatusCode.InternalServerError);
         }
 
         [Test]
@@ -146,7 +146,7 @@ namespace NClient.Tests.ClientTests
             result.Should().NotBeNull();
             using var assertionScope = new AssertionScope();
             result.Data.Should().Be(id);
-            result.StatusCode.Should().Be(HttpStatusCode.OK);
+            result.StatusCode.Should().Be((int)HttpStatusCode.OK);
             result.Error.Should().BeNull();
         }
 
@@ -173,7 +173,7 @@ namespace NClient.Tests.ClientTests
 
             result.Should().NotBeNull();
             using var assertionScope = new AssertionScope();
-            result.StatusCode.Should().Be(HttpStatusCode.BadRequest);
+            result.StatusCode.Should().Be((int)HttpStatusCode.BadRequest);
             result.Error.Should().BeEquivalentTo(BadRequestError);
         }
 
@@ -187,7 +187,7 @@ namespace NClient.Tests.ClientTests
 
             result.Should().NotBeNull();
             using var assertionScope = new AssertionScope();
-            result.StatusCode.Should().Be(HttpStatusCode.InternalServerError);
+            result.StatusCode.Should().Be((int)HttpStatusCode.InternalServerError);
             result.Error.Should().BeEquivalentTo(InternalServerError);
         }
 
@@ -252,7 +252,7 @@ namespace NClient.Tests.ClientTests
 
             result.Should().NotBeNull();
             using var assertionScope = new AssertionScope();
-            result.StatusCode.Should().Be(HttpStatusCode.OK);
+            result.StatusCode.Should().Be((int)HttpStatusCode.OK);
         }
 
         [Test]
@@ -265,7 +265,7 @@ namespace NClient.Tests.ClientTests
 
             result.Should().NotBeNull();
             using var assertionScope = new AssertionScope();
-            result.StatusCode.Should().Be(HttpStatusCode.BadRequest);
+            result.StatusCode.Should().Be((int)HttpStatusCode.BadRequest);
         }
 
         [Test]
@@ -278,7 +278,7 @@ namespace NClient.Tests.ClientTests
 
             result.Should().NotBeNull();
             using var assertionScope = new AssertionScope();
-            result.StatusCode.Should().Be(HttpStatusCode.BadRequest);
+            result.StatusCode.Should().Be((int)HttpStatusCode.BadRequest);
             result.Content.ToString().Should().BeEquivalentTo(JsonSerializer.Serialize(BadRequestError));
         }
 
@@ -292,7 +292,7 @@ namespace NClient.Tests.ClientTests
 
             result.Should().NotBeNull();
             using var assertionScope = new AssertionScope();
-            result.StatusCode.Should().Be(HttpStatusCode.InternalServerError);
+            result.StatusCode.Should().Be((int)HttpStatusCode.InternalServerError);
         }
 
         [Test]
@@ -305,7 +305,7 @@ namespace NClient.Tests.ClientTests
 
             result.Should().NotBeNull();
             using var assertionScope = new AssertionScope();
-            result.StatusCode.Should().Be(HttpStatusCode.OK);
+            result.StatusCode.Should().Be((int)HttpStatusCode.OK);
             result.Error.Should().BeNull();
         }
 
@@ -332,7 +332,7 @@ namespace NClient.Tests.ClientTests
 
             result.Should().NotBeNull();
             using var assertionScope = new AssertionScope();
-            result.StatusCode.Should().Be(HttpStatusCode.BadRequest);
+            result.StatusCode.Should().Be((int)HttpStatusCode.BadRequest);
             result.Error.Should().BeEquivalentTo(BadRequestError);
         }
 
@@ -346,7 +346,7 @@ namespace NClient.Tests.ClientTests
 
             result.Should().NotBeNull();
             using var assertionScope = new AssertionScope();
-            result.StatusCode.Should().Be(HttpStatusCode.InternalServerError);
+            result.StatusCode.Should().Be((int)HttpStatusCode.InternalServerError);
             result.Error.Should().BeEquivalentTo(InternalServerError);
         }
     }
